Navbar updated for user based

+++ src/app/features/login/services/auth.service.ts
@@ -1,7 +1,8 @@
 import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { Login } from '../models/login.model';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 import { BehaviorSubject, Observable } from 'rxjs';
+import { Login } from '../models/login.model';
 import { LoginResponse } from '../models/login-response.model';
 import { environment } from '../../../../environments/environment';
 import { User } from '../models/user.model';
@@ -14,7 +15,11 @@ export class AuthService {
 
   $user = new BehaviorSubject<User | undefined>(undefined);
 
-  constructor(private http: HttpClient, private cookieService: CookieService) { }
+  constructor(
+    private http: HttpClient,
+    private cookieService: CookieService,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) { }
 
   login(request: Login): Observable<LoginResponse> {
     return this.http.post<LoginResponse>(`${environment.apiBaseUrl}/api/Auth/login`, {
@@ -25,8 +30,11 @@ export class AuthService {
 
   setUser(user: User): void {
     this.$user.next(user);
-    localStorage.setItem('user-email', user.email)
-    localStorage.setItem('user-roles', user.roles.join(','))
+
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem('user-email', user.email);
+      localStorage.setItem('user-roles', user.roles.join(','));
+    }
   }
 
   user(): Observable<User | undefined> {
@@ -34,21 +42,24 @@ export class AuthService {
   }
 
   getUser(): User | undefined {
-    const email = localStorage.getItem('user-email');
-    const roles = localStorage.getItem('user-roles');
+    if (isPlatformBrowser(this.platformId)) {
+      const email = localStorage.getItem('user-email');
+      const roles = localStorage.getItem('user-roles');
 
-    if (email && roles) {
-      const user: User = {
-        email: email,
-        roles: roles.split(',')
-      };
-      return user;
+      if (email && roles) {
+        return {
+          email: email,
+          roles: roles.split(',')
+        } as User;
+      }
     }
     return undefined;
   }
 
   logout(): void {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
     this.cookieService.delete('Authorization', '/');
     this.$user.next(undefined);
   }
